refactor(server): separate station query and analytics flows
