+++ apps/server/src/domain/stations/repository/read/station.read.repository.ts
@@ -12,9 +12,6 @@ import type {
   NearestSearchArgs,
   NearestStationRow,
   StationContextRow,
-  StationRevenueGroupBy,
-  StationRevenuePoint,
-  StationRevenueRow,
   StationWorkerRow,
 } from "../../models";
 import type { StationQueryRepo } from "../station.repository.types";
@@ -45,26 +42,7 @@ export type StationReadRepo = Pick<
   | "findIdNameAddressByIds"
   | "listContextExcludingId"
   | "listNearest"
-  | "getRevenueByStation"
-  | "getRevenueForStation"
-  | "getRevenueSeries"
 >;
-
-function toRevenueBucketDate(value: Date, groupBy: StationRevenueGroupBy): Date {
-  const year = value.getUTCFullYear();
-  const month = value.getUTCMonth();
-  const day = value.getUTCDate();
-
-  if (groupBy === "YEAR") {
-    return new Date(Date.UTC(year, 0, 1, 0, 0, 0, 0));
-  }
-
-  if (groupBy === "MONTH") {
-    return new Date(Date.UTC(year, month, 1, 0, 0, 0, 0));
-  }
-
-  return new Date(Date.UTC(year, month, day, 0, 0, 0, 0));
-}
 
 export function makeStationReadRepository(
   client: PrismaClient | PrismaTypes.TransactionClient,
@@ -465,161 +443,5 @@ export function makeStationReadRepository(
       }).pipe(defectOn(StationRepositoryError));
     },
 
-    getRevenueByStation({ from, to }) {
-      return Effect.tryPromise({
-        try: async () => {
-          const rows = await client.rental.groupBy({
-            by: ["startStationId"],
-            where: {
-              status: "COMPLETED",
-              endTime: {
-                gte: from,
-                lte: to,
-              },
-            },
-            _count: { _all: true },
-            _sum: {
-              totalPrice: true,
-              duration: true,
-            },
-            _avg: {
-              duration: true,
-            },
-          });
-
-          const stationIds = rows.map(row => row.startStationId);
-          const stations = stationIds.length === 0
-            ? []
-            : await client.station.findMany({
-                where: { id: { in: stationIds } },
-                select: {
-                  id: true,
-                  name: true,
-                  address: true,
-                },
-              });
-          const stationMap = new Map(stations.map(station => [station.id, station]));
-
-          return rows.flatMap((row): StationRevenueRow[] => {
-            const station = stationMap.get(row.startStationId);
-            if (!station) {
-              return [];
-            }
-
-            return [{
-              stationId: station.id,
-              name: station.name,
-              address: station.address,
-              totalRentals: row._count._all,
-              totalRevenue: row._sum.totalPrice === null ? 0 : Number(row._sum.totalPrice),
-              totalDuration: row._sum.duration ?? 0,
-              avgDuration: row._avg.duration === null ? 0 : Number(row._avg.duration.toFixed(2)),
-            }];
-          });
-        },
-        catch: cause =>
-          new StationRepositoryError({
-            operation: "getRevenueByStation",
-            cause,
-          }),
-      }).pipe(defectOn(StationRepositoryError));
-    },
-
-    getRevenueForStation({ stationId, from, to }) {
-      return Effect.tryPromise({
-        try: async () => {
-          const aggregate = await client.rental.aggregate({
-            where: {
-              status: "COMPLETED",
-              startStationId: stationId,
-              endTime: {
-                gte: from,
-                lte: to,
-              },
-            },
-            _count: { _all: true },
-            _sum: {
-              totalPrice: true,
-              duration: true,
-            },
-            _avg: {
-              duration: true,
-            },
-          });
-
-          if (aggregate._count._all === 0) {
-            return null;
-          }
-
-          return {
-            totalRentals: aggregate._count._all,
-            totalRevenue: aggregate._sum.totalPrice === null ? 0 : Number(aggregate._sum.totalPrice),
-            totalDuration: aggregate._sum.duration ?? 0,
-            avgDuration: aggregate._avg.duration === null ? 0 : Number(aggregate._avg.duration.toFixed(2)),
-          };
-        },
-        catch: cause =>
-          new StationRepositoryError({
-            operation: "getRevenueForStation",
-            cause,
-          }),
-      }).pipe(defectOn(StationRepositoryError));
-    },
-
-    getRevenueSeries({ from, to, groupBy, stationId }) {
-      return Effect.tryPromise({
-        try: async () => {
-          const rows = await client.rental.findMany({
-            where: {
-              status: "COMPLETED",
-              ...(stationId ? { startStationId: stationId } : {}),
-              endTime: {
-                gte: from,
-                lte: to,
-              },
-            },
-            select: {
-              endTime: true,
-              totalPrice: true,
-            },
-          });
-
-          const buckets = new Map<string, StationRevenuePoint>();
-
-          for (const row of rows) {
-            if (!row.endTime) {
-              continue;
-            }
-
-            const bucketDate = toRevenueBucketDate(row.endTime, groupBy);
-            const key = bucketDate.toISOString();
-            const amount = row.totalPrice === null ? 0 : Number(row.totalPrice);
-            const current = buckets.get(key);
-
-            if (current) {
-              buckets.set(key, {
-                ...current,
-                totalRevenue: current.totalRevenue + amount,
-                totalRentals: current.totalRentals + 1,
-              });
-              continue;
-            }
-
-            buckets.set(key, {
-              date: bucketDate,
-              totalRevenue: amount,
-              totalRentals: 1,
-            });
-          }
-
-          return [...buckets.values()].sort((left, right) => left.date.getTime() - right.date.getTime());
-        },
-        catch: cause =>
-          new StationRepositoryError({
-            operation: "getRevenueSeries",
-            cause,
-          }),
-      }).pipe(defectOn(StationRepositoryError));
-    },
   };
 }
