+++ apps/server/src/http/controllers/stations/public.controller.ts
@@ -3,7 +3,7 @@ import type { RouteHandler } from "@hono/zod-openapi";
 import { Effect, Match } from "effect";
 
 import { withLoggedCause } from "@/domain/shared";
-import { StationQueryServiceTag } from "@/domain/stations";
+import { StationQueryServiceTag, StationStatsServiceTag } from "@/domain/stations";
 import {
   toContractNearbyStation,
   toContractStationReadSummary,
@@ -150,7 +150,7 @@ const getAllStationsRevenue: RouteHandler<StationsRoutes["getAllStationsRevenue"
 
   const eff = withLoggedCause(
     Effect.gen(function* () {
-      const service = yield* StationQueryServiceTag;
+      const service = yield* StationStatsServiceTag;
       return yield* service.getRevenueByStation({
         ...rangeResult.range,
         groupBy: rangeResult.groupBy,
