+++ apps/server/src/http/shared/features/station.layers.ts
@@ -1,10 +1,12 @@
 import { Layer } from "effect";
 
 import {
+  StationAnalyticsRepositoryLive,
   StationCommandRepositoryLive,
   StationCommandServiceLive,
   StationQueryRepositoryLive,
   StationQueryServiceLive,
+  StationStatsServiceLive,
 } from "@/domain/stations";
 
 import { PrismaLive } from "../infra.layers";
@@ -18,6 +20,10 @@ export const StationCommandReposLive = StationCommandRepositoryLive.pipe(
   Layer.provide(PrismaLive),
 );
 
+export const StationAnalyticsReposLive = StationAnalyticsRepositoryLive.pipe(
+  Layer.provide(PrismaLive),
+);
+
 export const StationQueryServiceLayer = StationQueryServiceLive.pipe(
   Layer.provide(StationQueryReposLive),
 );
@@ -26,10 +32,16 @@ export const StationCommandServiceLayer = StationCommandServiceLive.pipe(
   Layer.provide(Layer.mergeAll(StationCommandReposLive, StationQueryReposLive, AgencyReposLive)),
 );
 
+export const StationStatsServiceLayer = StationStatsServiceLive.pipe(
+  Layer.provide(Layer.mergeAll(StationAnalyticsReposLive, StationQueryReposLive)),
+);
+
 export const StationDepsLive = Layer.mergeAll(
+  StationAnalyticsReposLive,
   StationQueryReposLive,
   StationCommandReposLive,
   StationQueryServiceLayer,
   StationCommandServiceLayer,
+  StationStatsServiceLayer,
   PrismaLive,
 );
